Add GET '/contacts/new' route
We are moving the error one step at a time.
Using some of the nomenclature associated to Rails, but with a twist.
The new action can't be named new(), because that's how objects get
instantiated, and it looks like the system gets confused. So we go with
the best next choice, the more verbose newContact():
  - The name of the controller is the plural of the resource: Contacts
  - The name of the GET action for creation: newContact